+++ src/sophronia/components/tables/distributor-films-table.tsx
@@ -1,4 +1,4 @@
-import { useDistributorFilms } from 'lib/distributors';
+import { useDistributorFilms } from 'lib/fetch/distributors';
 import { FilmTable } from 'components/tables/film-table';
 import { Pagination } from 'components/ui/pagination';
 import { paginate } from 'lib/utils/pagination';
